requirements.txt
async-timeout==4.0.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
filelock==3.13.1
Flask==3.0.2
fonttools==4.50.0
fsspec==2024.3.1
huggingface-hub==0.21.4
idna==3.6
imageio==2.34.0
importlib_metadata==7.0.2
importlib_resources==6.3.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lime==0.2.0.1
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
packaging==24.0
pillow==10.2.0
plotly==5.20.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyWavelets==1.4.1
PyYAML==6.0.1
redis==5.0.3
regex==2023.12.25
requests==2.31.0
retrying==1.3.4
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
spotipy==2.23.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tifffile==2023.7.10
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.2
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.18.1
gunicorn==21.2.0
